Move documentation and package handlers out of router
